+++ src/utils/attributesGenerator.js
